+++ src/commands/advanced_generate.ts
@@ -311,6 +311,13 @@ export default class extends Command {
         prompt = style.prompt.slice().replace("{p}", prompt)
         prompt = prompt.replace("{np}", !negative_prompt || prompt.includes("###") ? negative_prompt : `###${negative_prompt}`)
         
+        if(party && party.wordlist.length) {
+            if(ctx.client.config.advanced?.dev) {
+                console.log(party.wordlist)
+            }
+            if(!party.wordlist.every(w => prompt.toLowerCase().includes(w))) return ctx.error({error: "Your prompt does not include all required words"})
+        }
+        
         if(ctx.client.config.advanced?.dev) {
             console.log(img?.height)
             console.log(img?.width)
@@ -318,7 +325,7 @@ export default class extends Command {
             console.log(width)
         }
 
-        const token = user_token || ctx.client.config.default_token || "0000000000"
+        const token = party?.shared_key || user_token || ctx.client.config.default_token || "0000000000"
         let img_data: Buffer | undefined
         if(img) {
             let img_data_res = await Centra(img.url, "GET")