+++ src/commands/end_party.ts
@@ -1,6 +1,7 @@
 import { SlashCommandBuilder } from "discord.js";
 import { Command } from "../classes/command";
 import { CommandContext } from "../classes/commandContext";
+import { SharedKeyDetails } from "@zeldafan0225/ai_horde";
 
 const command_data = new SlashCommandBuilder()
     .setName("end_party")
@@ -28,9 +29,18 @@ export default class extends Command {
         if(!party_data?.rowCount) return ctx.error({error: "Unable to end party"})
         ctx.client.cache.delete(`party-${ctx.interaction.channelId}`)
 
+        let usagestats: SharedKeyDetails = {}
+
+        if(party.shared_key) {
+            const usertoken = await ctx.client.getUserToken(ctx.interaction.user.id, ctx.database)
+            usagestats = await ctx.ai_horde_manager.getSharedKey(party.shared_key, {token: usertoken}).catch(console.error) || {}
+
+            await ctx.ai_horde_manager.deleteSharedKey(party.shared_key, {token: usertoken}).catch(console.error)
+        }
+
         await ctx.interaction.reply({content: "Party ended.", ephemeral: true})
         ctx.interaction.channel?.send({
-            content: `The party police showed up and broke down this party.\n${party_data.rows[0].users?.length} users participated.\nThanks to <@${party.creator_id}> for hosting this party`
+            content: `The party police showed up and broke down this party.\n${party_data.rows[0].users?.length} users participated.${usagestats ? `\n${usagestats.utilized} kudos have been spent by <@${party.creator_id}> only for generations in this party` : ""}\nThanks to <@${party.creator_id}> for hosting this party`
         })
     }
 }