+++ src/commands/party.ts
@@ -48,6 +48,16 @@ const command_data = new SlashCommandBuilder()
             .setName("recurring")
             .setDescription("If users get rewarded for each generation or only their first")
         )
+        .addBooleanOption(
+            new SlashCommandBooleanOption()
+            .setName("pay_for_generations")
+            .setDescription("Whether to pay for the generations users make")
+        )
+        .addStringOption(
+            new SlashCommandStringOption()
+            .setName("wordlist")
+            .setDescription("Set a comma separated list of words the users prompt has to include")
+        )
     }
 
 
@@ -69,6 +79,8 @@ export default class extends Command {
         const award = ctx.interaction.options.getInteger("award", true)
         const duration = ctx.interaction.options.getInteger("duration", true)
         const recurring = !!(ctx.interaction.options.getBoolean("recurring") ?? ctx.client.config.party?.default?.recurring)
+        const pay = !!(ctx.interaction.options.getBoolean("pay_for_generations") ?? ctx.client.config.party?.default?.pay_for_generations)
+        const wordlist = (ctx.interaction.options.getString("wordlist") ?? "").split(",").map(w => w.trim().toLowerCase())
         const style_raw = ctx.interaction.options.getString("style") ?? ctx.client.config.generate?.default?.style ?? "raw"
         const style = ctx.client.horde_styles[style_raw.toLowerCase()]
 
@@ -82,6 +94,13 @@ export default class extends Command {
         if(!style) return ctx.error({error: "A valid style is required"})
         if(ctx.client.config.generate?.blacklisted_styles?.includes(style_raw.toLowerCase())) return ctx.error({error: "The chosen style is blacklisted"})
 
+        if(ctx.client.config.party.user_restrictions?.wordlist) {
+            if(
+                ctx.client.config.party.user_restrictions?.wordlist.min || 0 > wordlist.length ||
+                ctx.client.config.party.user_restrictions?.wordlist.max && ctx.client.config.party.user_restrictions?.wordlist.max < wordlist.length
+            ) return ctx.error({error: `Your wordlist must be between ${ctx.client.config.party.user_restrictions?.wordlist.min || "no"} and ${ctx.client.config.party.user_restrictions?.wordlist.max || "unlimited"} words`})
+        }
+
         await ctx.interaction.deferReply({ephemeral: true})
 
         const thread = await ctx.interaction.channel.threads.create({
@@ -90,13 +109,30 @@ export default class extends Command {
         }).catch(console.error)
         if(!thread?.id) return ctx.error({error: "Unable to start party"})
 
-        const party = await ctx.database.query(`INSERT INTO parties (channel_id, guild_id, creator_id, ends_at, style, award, recurring) VALUES ($1, $2, $3, CURRENT_TIMESTAMP + interval '${duration} day', $4, $5, $6) RETURNING *`, [
+        let shared_key_id: string | null = null
+
+        if(pay) {
+            const token = await ctx.client.getUserToken(ctx.interaction.user.id, ctx.database)
+
+            const shared_key = await ctx.ai_horde_manager.putSharedKey({
+                kudos: 100000,
+                expiry: duration,
+                name: `Party ${name}`
+            }, {token}).catch(console.error)
+
+            if(shared_key?.id) shared_key_id = shared_key.id
+            if(ctx.client.config.advanced?.dev) console.log(shared_key_id)
+        }
+
+        const party = await ctx.database.query(`INSERT INTO parties (channel_id, guild_id, creator_id, ends_at, style, award, recurring, shared_key, wordlist) VALUES ($1, $2, $3, CURRENT_TIMESTAMP + interval '${duration} day', $4, $5, $6, $7, $8) RETURNING *`, [
             thread.id,
             thread.guildId,
             ctx.interaction.user.id,
             style_raw.toLowerCase(),
             award,
-            recurring
+            recurring,
+            shared_key_id,
+            wordlist
         ]).catch(console.error)
 
         if(!party?.rowCount) {
@@ -105,7 +141,11 @@ export default class extends Command {
         }
 
         const start = await thread.send({
-            content: `<@${ctx.interaction.user.id}> started the party "${name}" with the style "${style_raw}".\nYou will get ${award} kudos for ${recurring ? `every generation` : `your first generation`}.\nThe party ends <t:${Math.round((Date.now() + 1000 * 60 * 60 * 24 * duration)/1000)}:R>\n\n${ctx.client.config.party.mention_roles?.length ? ctx.client.config.party.mention_roles.map(r => `<@&${r}>`).join(" ") : ""}`
+            content: `<@${ctx.interaction.user.id}> started the party "${name}" with the style "${style_raw}".\nYou will get ${award} kudos for ${recurring ? `every generation` : `your first generation`}.\nThe party ends <t:${Math.round((Date.now() + 1000 * 60 * 60 * 24 * duration)/1000)}:R>${wordlist.length ? `\nThe prompt has to include the words: ${wordlist.join(",")}` : ""}${pay && shared_key_id ? "\nThe party creator will pay for all generations 🥳" : ""}\n\n${ctx.client.config.party.mention_roles?.length ? ctx.client.config.party.mention_roles.map(r => `<@&${r}>`).join(" ") : ""}`,
+            allowedMentions: {
+                users: [ctx.interaction.user.id],
+                roles: ctx.client.config.party.mention_roles
+            }
         }).catch(console.error)
 
         await start?.pin().catch(console.error)