+++ src/index.ts
@@ -6,7 +6,7 @@ import { handleComponents } from "./handlers/componentHandler";
 import { handleModals } from "./handlers/modalHandler";
 import { Pool } from "pg"
 import { handleAutocomplete } from "./handlers/autocompleteHandler";
-import {AIHorde} from "@zeldafan0225/ai_horde";
+import { AIHorde } from "@zeldafan0225/ai_horde";
 import { handleContexts } from "./handlers/contextHandler";
 import {existsSync, mkdirSync} from "fs"
 import { handleMessageReact } from "./handlers/messageReact";
@@ -41,7 +41,7 @@ if(client.config.use_database !== false) {
     
     connection.connect().then(async () => {
         await connection!.query("CREATE TABLE IF NOT EXISTS user_tokens (index SERIAL, id VARCHAR(100) PRIMARY KEY, token VARCHAR(100) NOT NULL, horde_id int NOT NULL DEFAULT 0)")
-        await connection!.query("CREATE TABLE IF NOT EXISTS parties (index SERIAL, channel_id VARCHAR(100) PRIMARY KEY, guild_id VARCHAR(100) NOT NULL, creator_id VARCHAR(100) NOT NULL, created_at TIMESTAMP NOT NULL DEFAULT CURRENT_TIMESTAMP, ends_at TIMESTAMP NOT NULL DEFAULT CURRENT_TIMESTAMP, style VARCHAR(1000) NOT NULL, award INT NOT NULL DEFAULT 1, recurring BOOLEAN NOT NULL DEFAULT false, users VARCHAR(100)[] NOT NULL DEFAULT '{}')")
+        await connection!.query("CREATE TABLE IF NOT EXISTS parties (index SERIAL, channel_id VARCHAR(100) PRIMARY KEY, guild_id VARCHAR(100) NOT NULL, creator_id VARCHAR(100) NOT NULL, created_at TIMESTAMP NOT NULL DEFAULT CURRENT_TIMESTAMP, ends_at TIMESTAMP NOT NULL DEFAULT CURRENT_TIMESTAMP, style VARCHAR(1000) NOT NULL, award INT NOT NULL DEFAULT 1, recurring BOOLEAN NOT NULL DEFAULT false, users VARCHAR(100)[] NOT NULL DEFAULT '{}', shared_key VARCHAR(100), wordlist text[] NOT NULL DEFAULT '{}')")
         await connection!.query("CREATE TABLE IF NOT EXISTS pending_kudos (index SERIAL, unique_id VARCHAR(200) PRIMARY KEY, target_id VARCHAR(100) NOT NULL, from_id VARCHAR(100) NOT NULL, amount int NOT NULL, updated_at TIMESTAMP NOT NULL DEFAULT CURRENT_TIMESTAMP)")
     }).catch(console.error);
 
@@ -86,8 +86,8 @@ client.on("ready", async () => {
     if(client.config.party?.enabled && !client.config.generate?.enabled) throw new Error("When party is enabled the /generate command also needs to be enabled")
 
     if(client.config.party?.enabled && connection) {
-        await client.cleanUpParties(connection)
-        setInterval(async () => await client.cleanUpParties(connection), 1000 * 60 * 5)
+        await client.cleanUpParties(ai_horde_manager, connection)
+        setInterval(async () => await client.cleanUpParties(ai_horde_manager, connection), 1000 * 60 * 5)
     }
 })